react query

+++ src/services/apiCabins.js
@@ -2,6 +2,7 @@ import supabase, { supabaseUrl } from "./supabase";
 
 export async function getCabins() {
   const { data, error } = await supabase.from("cabins").select("*");
+
   if (error) {
     console.error(error);
     throw new Error("Cabins could not be loaded");
@@ -10,83 +11,46 @@ export async function getCabins() {
   return data;
 }
 
-// export async function createEditCabin(newCabin, id) {
-//   const hasImagePath = newCabin.image?.startsWith?.(supabaseUrl);
+export async function createEditCabin(newCabin, id) {
+  const hasImagePath = newCabin.image?.startsWith?.(supabaseUrl);
 
-//   const imageName = `${Math.random()}-${newCabin.image.name}`.replaceAll(
-//     "/",
-//     ""
-//   );
-//   const imagePath = hasImagePath
-//     ? newCabin.image
-//     : `${supabaseUrl}/storage/v1/object/public/cabin-images/${imageName}`;
-
-//   // 1. Create/edit cabin
-//   let query = supabase.from("cabins");
-
-//   // A) CREATE
-//   if (!id) query = query.insert([{ ...newCabin, image: imagePath }]);
-
-//   // B) EDIT
-//   if (id) query = query.update({ ...newCabin, image: imagePath }).eq("id", id);
-
-//   const { data, error } = await query.select().single();
-
-//   if (error) {
-//     console.error(error);
-//     throw new Error("Cabin could not be created");
-//   }
-
-//   // 2. Upload image
-//   if (hasImagePath) return data;
-
-//   const { error: storageError } = await supabase.storage
-//     .from("cabin-images")
-//     .upload(imageName, newCabin.image);
-
-//   // 3. Delete the cabin IF there was an error uplaoding image
-//   if (storageError) {
-//     await supabase.from("cabins").delete().eq("id", data.id);
-//     console.error(storageError);
-//     throw new Error(
-//       "Cabin image could not be uploaded and the cabin was not created"
-//     );
-//   }
-
-//   return data;
-// }
-
-export async function createCabin(newCabin) {
   const imageName = `${Math.random()}-${newCabin.image.name}`.replaceAll(
     "/",
     ""
   );
+  const imagePath = hasImagePath
+    ? newCabin.image
+    : `${supabaseUrl}/storage/v1/object/public/cabin-images/${imageName}`;
 
-  // https://dysfrqyevghjjcljztne.supabase.co/storage/v1/object/public/cabin-images/cabin-001.jpg
+  // 1. Create/edit cabin
+  let query = supabase.from("cabins");
 
-  const imagePath = `${supabaseUrl}/storage/v1/object/public/cabin-images/${imageName}`;
+  // A) CREATE
+  if (!id) query = query.insert([{ ...newCabin, image: imagePath }]);
 
-  const { data, error } = await supabase
-    .from("cabins")
-    .insert([{ ...newCabin, image: imagePath }]);
+  // B) EDIT
+  if (id) query = query.update({ ...newCabin, image: imagePath }).eq("id", id);
+
+  const { data, error } = await query.select().single();
 
   if (error) {
-    console.log(error);
+    console.error(error);
     throw new Error("Cabin could not be created");
   }
 
-  //2. upload image
+  // 2. Upload image
+  if (hasImagePath) return data;
 
   const { error: storageError } = await supabase.storage
     .from("cabin-images")
     .upload(imageName, newCabin.image);
 
-  //3. delete the cabin if there was an error uploading the image
+  // 3. Delete the cabin IF there was an error uplaoding image
   if (storageError) {
     await supabase.from("cabins").delete().eq("id", data.id);
     console.error(storageError);
     throw new Error(
-      "Cabins image could not be uploaded and the cabin was not created"
+      "Cabin image could not be uploaded and the cabin was not created"
     );
   }
 
@@ -95,9 +59,10 @@ export async function createCabin(newCabin) {
 
 export async function deleteCabin(id) {
   const { data, error } = await supabase.from("cabins").delete().eq("id", id);
+
   if (error) {
     console.error(error);
-    throw new Error("Cabins could not be deleted");
+    throw new Error("Cabin could not be deleted");
   }
 
   return data;
